requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Data fetching
requests>=2.25.0
urllib3>=1.26.0
netCDF4>=1.5.7
xarray>=0.19.0

# Weather data processing
metpy>=1.3.0
cartopy>=0.20.0
pyproj>=3.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=0.19.0

# Utilities
tqdm>=4.62.0
click>=8.0.0
pathlib2>=2.3.6

# Development dependencies
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.0.0
isort>=5.10.0
flake8>=3.9.0
mypy>=0.910
bandit>=1.7.0
pylint>=2.12.0
pre-commit>=2.15.0

# Type stubs for better type checking
pandas-stubs>=2.0.0
types-requests>=2.25.0
types-PyYAML>=6.0.0
types-tqdm>=4.62.0
types-click>=7.1.0

# Optional: For advanced plotting
plotly>=5.0.0
bokeh>=2.4.0

# Optional: For high-quality SVG weather symbols with transparency
# Install system dependencies first:
#   macOS: brew install cairo
#   Ubuntu/Debian: sudo apt-get install libcairo2-dev
# Then install these Python packages:
cairosvg>=2.5.0
pillow>=8.0.0

# Jupyter environment for interactive testing
jupyter>=1.0.0
jupyterlab>=3.0.0
ipywidgets>=7.6.0
ipympl>=0.8.0
